New hook for dynamicLoadingMessage
